FIX(backend): Map Kotlin enums to PostgreSQL native enum columns. EN: This change keeps Spring Data queries compatible with the existing PostgreSQL enum schema so public and admin endpoints can run without SQL type errors. ES: Este cambio mantiene las consultas de Spring Data compatibles con los enums nativos de PostgreSQL para que los endpoints publicos y administrativos funcionen sin errores de tipo SQL.

+++ backend/src/main/kotlin/com/api/travel_api/user/UserRepository.kt
@@ -3,6 +3,6 @@ package com.api.travel_api.user
 import com.api.travel_api.model.entities.User
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface UserRepository : JpaRepository<User, Long> {
+interface UserRepository : JpaRepository<User, Int> {
     fun findByEmail(email: String): User?
 }
